requirements.txt
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
Babel==2.6.0
beautifulsoup4==4.7.1
certifi==2018.11.29
css-html-js-minify==2.5.5
csvkit==1.0.3
dbfread==2.0.7
dicttoxml==1.7.4
Django==2.1.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
isodate==0.6.0
jdcal==1.4
leather==0.3.3
lxml==4.3.1
Markdown==3.0.1
openpyxl==2.6.0
parsedatetime==2.4
Pillow==5.4.1
PyJWT==1.7.1
python-slugify==2.0.1
pytimeparse==1.1.8
pytz==2018.9
six==1.12.0
soupsieve==1.7.3
SQLAlchemy==1.2.17
Unidecode==1.0.23
wincertstore==0.2
xlrd==1.2.0
xmlutils==1.4
